+++ enterprise/api/v1/views/enterprise_customer_user.py
@@ -6,6 +6,7 @@ from rest_framework import filters
 
 from enterprise import models
 from enterprise.api.filters import EnterpriseCustomerUserFilterBackend
+from enterprise.api.pagination import PaginationWithFeatureFlags
 from enterprise.api.v1 import serializers
 from enterprise.api.v1.views.base_views import EnterpriseReadWriteModelViewSet
 
@@ -17,6 +18,7 @@ class EnterpriseCustomerUserViewSet(EnterpriseReadWriteModelViewSet):
 
     queryset = models.EnterpriseCustomerUser.objects.all()
     filter_backends = (filters.OrderingFilter, DjangoFilterBackend, EnterpriseCustomerUserFilterBackend)
+    pagination_class = PaginationWithFeatureFlags
 
     FIELDS = (
         'enterprise_customer', 'user_id', 'active',