+++ enterprise/models.py
@@ -4212,7 +4212,7 @@ class EnterpriseCustomerSsoConfiguration(TimeStampedModel, SoftDeletableModel):
         return sp_metadata_url
 
 
-class EnterpriseGroup(TimeStampedModel):
+class EnterpriseGroup(TimeStampedModel, SoftDeletableModel):
     """
     Enterprise Group model
 
@@ -4233,6 +4233,14 @@ class EnterpriseGroup(TimeStampedModel):
         related_name='groups',
         on_delete=models.deletion.CASCADE
     )
+    applies_to_all_contexts = models.BooleanField(
+        verbose_name="Set group membership to the entire org of learners.",
+        default=False,
+        help_text=_(
+            "When enabled, all learners connected to the org will be considered a member."
+        )
+    )
+
     history = HistoricalRecords()
 
     class Meta:
@@ -4242,7 +4250,7 @@ class EnterpriseGroup(TimeStampedModel):
         ordering = ['-modified']
 
 
-class EnterpriseGroupMembership(TimeStampedModel):
+class EnterpriseGroupMembership(TimeStampedModel, SoftDeletableModel):
     """
     Enterprise Group Membership model