+++ CHANGELOG.rst
@@ -15,6 +15,25 @@ Change Log
 
 Unreleased
 ----------
+[4.13.3]
+---------
+* feat: adding management command to remove expired pending group memberships
+
+[4.13.2]
+---------
+* feat: add a waffle flag for enterprise groups feature
+
+[4.13.1]
+---------
+* feat: adding soft delete functionality for groups and group memberships
+
+[4.13.0]
+---------
+* feat: add Waffle-based `enterprise_features` to the `EnterpriseCustomerUserViewSet`.
+
+[4.12.6]
+---------
+* fix: Proximus learner transmission failures
 
 [4.12.6]
 ---------